requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
decorator==4.0.11
google-api-python-client==1.6.3
httplib2==0.10.3
idna==2.6
imageio==2.1.2
moviepy==0.2.3.2
numpy==1.13.1
oauth2client==4.1.2
oauthlib==2.0.2
olefile==0.44
Pillow==4.2.1
pyasn1==0.3.3
pyasn1-modules==0.1.1
pytube==6.4.2
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
six==1.10.0
tqdm==4.11.2
tweepy==3.5.0
uritemplate==3.0.0
urllib3==1.22
